Move settings and data files to .config and .local
Move the settings file to $HOME/.config/twitcheckrc.
When the script runs and does not find a settings file there (e.g. on first
run), it will create it from the default settings distributed with the
script.
If the data and db files do not exist, create their directories and them.