Bugfix: Don't falsely remove online status from db
In the following situation, a channel could be marked wrongly as offline:
If another channel whose name was the same as this channel's beginning was
in the checklist, but offline.

Example:
  happy
  happy_gamer

happy_gamer would always be removed whenever happy was offline, resulting in
constant notifications.

This commit fixes that bug.