Added option to sent twitch app URI instead of URL to pushbullet.
There is now a setting for the pushbullet module that allows you to send the URI instead of URL so that clicking on the notification on mobile devices with the app installed, will open the stream in the twitch app instead of the website. The raw channel name now also gets passed to modules as the 5th argument.