Add fix for using script through cron on OS X.
For whatever reason, on OS X, cron's env does not have '/usr/local/bin' in its PATH. This is where homebrew installs jq. So when added to crontab on OS X, the twitcheck.sh fails to find jq. We are simply appending PATH to add the directory. This is very platform specific and will probably break something in the future.