Attempt to handle bad API returns better
So far, we didn't handle bad API returns correctly. This commit attempts
a fix by removing the bad entries and manually re-inserting the data from
the database into the cached return, which in the end gets parsed and saved
to the database again. It wasn't fully tested, but should work just fine.