Rebult pushbullet module.
I wanted the pushbullet module to be able to send a URL to some targets and a URI to others. I have rebuilt the module to support a space seperated list for each (as defined in the twitcheckrc file). The module will check to see if there are values in either setting and send to those targets for each. If neither has values, it will send to all targets. I also changed the formatting of the notifications so that it shows more info on smaller screens.