replace sed workaround with `jq -R` invocation
This commit replaces the ugly sed workaround to encode double quotes in the
recovered json with a simply call to `jq -R '.'`. The -R argument for jq
causes it to parse the input data as raw text, and then output it as a json-
formatted string. By using this functionality of jq, we delegate the handling
of all corner cases to jq, which most certainly does a better job at this
than we could.