Add interactive streamer list
When running the script with the -i command line flag, show an updating
list of online streams by parsing the offline database. Any notifications
that appear will take effect within (at most) ten seconds. This feels like
a good tradeoff so we don't constantly write to the screen while nothing
changed.

Much credit to
https://stackoverflow.com/questions/27945567/bash-script-overwrite-output-without-filling-scrollback/27946484#27946484
for showing a way to update the screen.