Fixed main function to use the notifier functions.
Since the script no longer uses the external modules, these changes change the way the main function calles the notifiers. We no longer need to pass the config to the modules and we do not need to know the module directory. These changes may be a little buggy and may need further testing.