Move sanity check for twitch API into it's service plugin
There was a check in notifier_gateway() that made sure twitch did not respond
with broken null results. While, it could possibly work for other services, it
would have only output database into to STDOUT.

This needed to be moved because the way the twitch service made use of this info
was to capture notifier_gateway()'s output, which mean anything
notifier_gateway() tried to print to STDOUT was intercepted.

Since other services like hitbox, have their own sanity checks, it makes
sense to move this into twitch's plugin since it was the only one using it.